+++ app/assets/javascripts/oxalis/view/action-bar/tracing_actions_view.js
@@ -17,6 +17,8 @@ import { copyAnnotationToUserAccount, finishAnnotation } from "admin/admin_rest_
 import { location } from "libs/window";
 import type { OxalisState, RestrictionsAndSettingsType, TaskType } from "oxalis/store";
 import type { APIUserType, APITracingType } from "admin/api_flow_types";
+import { layoutEmitter } from "oxalis/view/layouting/layout_persistence";
+import { updateUserSettingAction } from "oxalis/model/actions/settings_actions";
 
 type StateProps = {
   tracingType: APITracingType,
@@ -231,6 +233,20 @@ class TracingActionsView extends PureComponent<StateProps, State> {
         />,
       );
     }
+
+    elements.push(
+      <Menu.Item key="reset-layout">
+        <div
+          onClick={() => {
+            Store.dispatch(updateUserSettingAction("layoutScaleValue", 1));
+            layoutEmitter.emit("resetLayout");
+          }}
+        >
+          <Icon type="laptop" />
+          Reset Layout
+        </div>
+      </Menu.Item>,
+    );
 
     const menu = <Menu>{elements}</Menu>;